cancer type column is no longer lazy loaded (fix https://github.com/cBioPortal/cbioportal/issues/2811)
