+++ src/pages/resultsView/mutation/Mutations.tsx
@@ -65,7 +65,6 @@ export default class Mutations extends React.Component<IMutationsPageProps, {}>
                     <MSKTab key={gene} id={gene} linkText={gene} anchorStyle={anchorStyle}>
                         <MutationMapper
                             studyId={this.props.store.studyId}
-                            studyToCancerType={this.props.store.studyToCancerType}
                             store={mutationMapperStore}
                             discreteCNACache={this.props.store.discreteCNACache}
                             oncoKbEvidenceCache={this.props.store.oncoKbEvidenceCache}
