+++ src/shared/components/mutationTable/column/CancerTypeColumnFormatter.tsx
@@ -1,90 +1,59 @@
 import * as React from "react";
-import LazyLoadedTableCell from "shared/lib/LazyLoadedTableCell";
-import {Mutation} from "../../../api/generated/CBioPortalAPI";
-import MutationTable, {IMutationTableProps} from "../MutationTable";
-import CancerTypeCache from "../../../cache/CancerTypeCache";
-import {CacheData} from "../../../lib/LazyMobXCache";
+import {Mutation} from "shared/api/generated/CBioPortalAPI";
+import TableCellStatusIndicator from "shared/components/TableCellStatus";
+import {TableCellStatus} from "shared/components/TableCellStatus";
 
 export default class CancerTypeColumnFormatter {
 
-    public static getData(d: Mutation[],
-                          studyId?: string,
-                          cancerTypeCache?: CancerTypeCache,
-                          studyCancerTypeMap?: {[studyId:string]:string}): CacheData<{value:string}>|null
+    public static getData(d: Mutation[], sampleIdToTumorType?: {[sampleId: string]: string}): string|null
     {
-        if (cancerTypeCache && studyId) {
-            let cacheDatum:CacheData<{value:string}>|null = cancerTypeCache.get({
-                entityId: d[0].sampleId,
-                studyId
-            });
+        let data: string|null = null;
 
-            if (cacheDatum && cacheDatum.status === "complete" && cacheDatum.data === null) {
-                // If no clinical data, use study cancer type if we have it
-                if (studyCancerTypeMap) {
-                    const cancerType = studyCancerTypeMap[studyId];
-                    if (cancerType) {
-                        cacheDatum = {
-                            status: "complete",
-                            data: {
-                                value: cancerType
-                            }
-                        };
-                    }
-                }
-            }
-            return cacheDatum;
-        } else {
-            return {
-                status: "error",
-                data: null
-            };
+        if (sampleIdToTumorType) {
+            data = sampleIdToTumorType[d[0].sampleId] || null;
         }
+
+        return data;
     }
 
-    public static makeRenderFunction<P extends IMutationTableProps>(table:MutationTable<P>) {
-        return LazyLoadedTableCell<Mutation[], {value:string}>(
-            (d:Mutation[])=>{
-                const cancerTypeCache:CancerTypeCache|undefined = table.props.cancerTypeCache;
-                const studyId:string|undefined = table.props.studyId;
-                const studyCancerTypeMap:{[studyId:string]:string} | undefined = table.props.studyToCancerType;
+    public static sortBy(d: Mutation[], sampleIdToTumorType?: {[sampleId: string]: string}): string|null
+    {
+        const data = CancerTypeColumnFormatter.getData(d, sampleIdToTumorType);
 
-                return CancerTypeColumnFormatter.getData(d, studyId, cancerTypeCache, studyCancerTypeMap);
-            },
-            (t:{value:string})=>(<span>{t.value}</span>),
-            "Cancer type not available for this sample."
-        );
-    }
-
-    public static sortBy(d:Mutation[], studyId?:string, cancerTypeCache?:CancerTypeCache) {
-        let ret;
-        if (cancerTypeCache && studyId) {
-            const cacheDatum = cancerTypeCache.get({
-                entityId:d[0].sampleId,
-                studyId: studyId
-            });
-            if (cacheDatum && cacheDatum.data) {
-                ret = cacheDatum.data.value;
-            } else {
-                ret = null;
-            }
-        } else {
-            ret = null;
+        if (data) {
+            return data;
         }
-        return ret;
+        else {
+            return null;
+        }
     }
 
     public static filter(d: Mutation[],
                          filterStringUpper: string,
-                         studyId?: string,
-                         cancerTypeCache?: CancerTypeCache,
-                         studyCancerTypeMap?: {[studyId:string]:string}): boolean
+                         sampleIdToTumorType?: {[sampleId: string]: string}): boolean
     {
-        const cacheDatum = CancerTypeColumnFormatter.getData(d, studyId, cancerTypeCache, studyCancerTypeMap);
+        const data = CancerTypeColumnFormatter.getData(d, sampleIdToTumorType);
 
         return (
-            cacheDatum !== null &&
-            cacheDatum.data !== null &&
-            cacheDatum.data.value.toUpperCase().indexOf(filterStringUpper) > -1
+            data !== null &&
+            data.toUpperCase().indexOf(filterStringUpper) > -1
         );
+    }
+
+    public static render(d: Mutation[], sampleIdToTumorType?: {[sampleId: string]: string})
+    {
+        const data = CancerTypeColumnFormatter.getData(d, sampleIdToTumorType);
+
+        if (data) {
+            return <span>{data}</span>;
+        }
+        else {
+            return (
+                <TableCellStatusIndicator
+                    status={TableCellStatus.NA}
+                    naAlt="Cancer type not available for this sample."
+                />
+            );
+        }
     }
 }
